# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: ReqFlow Procurement Workflow API
servers:
  - url: /api
...
paths:
...
  /requisitions:
    get:
      operationId: listRequisitions
      tags: [requisitions]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: raised_by_id
          in: query
          schema:
            type: integer
        - name: assigned_to_id
          in: query
          schema:
            type: integer
        - name: checker1_id
          in: query
          schema:
            type: integer
        - name: checker2_id
          in: query
          schema:
            type: integer
        - name: approver_id
          in: query
          schema:
            type: integer
        - name: checker_id
          in: query
          schema:
            type: integer
        - name: project_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Matches ref number, requester, purpose, site, item names, or vessel/equipment/location names.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requisition"
    post:
      operationId: createRequisition
      tags: [requisitions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}:
    get:
      operationId: getRequisition
      tags: [requisitions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "404":
          description: Not found
    patch:
      operationId: updateRequisition
      tags: [requisitions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
    delete:
      operationId: deleteRequisition
      tags: [requisitions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /requisitions/{id}/items:
    get:
      operationId: listRequisitionItems
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequisitionItem"
    post:
      operationId: addRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionItem"
  /requisitions/{id}/items/{itemId}:
    patch:
      operationId: updateRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionItem"
    delete:
      operationId: deleteRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /requisitions/{id}/attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
...
components:
  schemas:
...
    Requisition:
      type: object
      required: [id, ref_number, raised_by_name, requisition_date, status, priority, created_at, updated_at]
      properties:
        id:
          type: integer
        ref_number:
          type: string
        project_id:
          type: ["integer", "null"]
        project_name:
          type: ["string", "null"]
        site_id:
          type: ["integer", "null"]
        site_name:
          type: ["string", "null"]
        raised_by_id:
          type: ["integer", "null"]
        raised_by_name:
          type: string
        requisition_date:
          type: string
        priority:
          type: string
        purpose:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        checker1_id:
          type: ["integer", "null"]
        checker1_name:
          type: ["string", "null"]
        checker1_suggestion:
          type: ["string", "null"]
        checker1_reviewed_at:
          type: ["string", "null"]
        checker1_status:
          type: ["string", "null"]
        checker2_id:
          type: ["integer", "null"]
        checker2_name:
          type: ["string", "null"]
        checker2_suggestion:
          type: ["string", "null"]
        checker2_reviewed_at:
          type: ["string", "null"]
        checker2_status:
          type: ["string", "null"]
        approver_id:
          type: ["integer", "null"]
        approver_name:
          type: ["string", "null"]
        approver_suggestion:
          type: ["string", "null"]
        approver_reviewed_at:
          type: ["string", "null"]
        approver_status:
          type: ["string", "null"]
        hold_reason:
          type: ["string", "null"]
        held_at:
          type: ["string", "null"]
        pre_hold_status:
          type: ["string", "null"]
        purchase_head_id:
          type: ["integer", "null"]
        purchase_head_name:
          type: ["string", "null"]
        assigned_to_id:
          type: ["integer", "null"]
        assigned_to_name:
          type: ["string", "null"]
        assigned_at:
          type: ["string", "null"]
        submitted_at:
          type: ["string", "null"]
        approved_at:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
        item_count:
          type: ["integer", "null"]
        total_expected_cost:
          type: ["number", "null"]
    RequisitionDetail:
      allOf:
        - $ref: "#/components/schemas/Requisition"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RequisitionItem"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/Attachment"
            queries:
              type: array
              items:
                $ref: "#/components/schemas/QueryWithReplies"
            status_updates:
              type: array
              items:
                $ref: "#/components/schemas/StatusUpdate"
            compliance:
              oneOf:
                - $ref: "#/components/schemas/ComplianceStatus"
                - type: "null"
...
    RequisitionInput:
      type: object
      required: [project_id, raised_by_name, requisition_date]
      properties:
        project_id:
          type: integer
        site_id:
          type: ["integer", "null"]
        raised_by_id:
          type: ["integer", "null"]
        raised_by_name:
          type: string
        site_name:
          type: string
        requisition_date:
          type: string
        priority:
          type: string
        purpose:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequisitionItemInput"
    RequisitionUpdate:
      type: object
      properties:
        project_id:
          type: ["integer", "null"]
        site_id:
          type: ["integer", "null"]
        raised_by_name:
          type: string
        site_name:
          type: string
        requisition_date:
          type: string
        priority:
          type: string
        purpose:
          type: string
        notes:
          type: string
    RequisitionItem:
      type: object
      required: [id, requisition_id, item_name, quantity]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        item_name:
          type: string
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        reference_no:
          type: ["string", "null"]
        expected_cost:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        asset_id:
          type: ["integer", "null"]
        asset_name:
          type: ["string", "null"]
        sort_order:
          type: integer
    RequisitionItemInput:
      type: object
      required: [item_name, quantity]
      properties:
        item_name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reference_no:
          type: string
        expected_cost:
          type: ["number", "null"]
        description:
          type: string
        remark:
          type: string
        asset_id:
          type: ["integer", "null"]
        asset_name:
          type: string
        sort_order:
          type: integer
    Attachment:
      type: object
      required: [id, requisition_id, filename, original_name]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        context:
          type: string
        context_id:
          type: ["integer", "null"]
        filename:
          type: string
        original_name:
          type: string
        mime_type:
          type: ["string", "null"]
        size_bytes:
          type: ["integer", "null"]
        uploaded_by_name:
          type: ["string", "null"]
        uploaded_at:
          type: string
        url:
          type: string
...
    QueryWithReplies:
      type: object
      required: [id, requisition_id, raised_by_name, message, is_resolved, created_at]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        raised_by_name:
          type: string
        raised_by_role:
          type: ["string", "null"]
        message:
          type: string
        is_resolved:
          type: boolean
        resolved_by_name:
          type: ["string", "null"]
        resolved_at:
          type: ["string", "null"]
        created_at:
          type: string
        replies:
          type: array
          items:
            $ref: "#/components/schemas/QueryReplyObject"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
...
    StatusUpdate:
      type: object
      required: [id, requisition_id, updated_by_name, stage, update_date, created_at]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        updated_by_name:
          type: string
        stage:
          type: string
        notes:
          type: ["string", "null"]
        update_date:
          type: string
        created_at:
          type: string
...